+++ client/src/components/Header/Header.tsx
@@ -1,20 +1,13 @@
 import { useState } from "react";
 import { Link } from "react-router";
-import { useAuth } from "../../services/AuthContext";
 import "./Header.css";
+import UserMenu from "./UserMenu/UserMenu";
 
 function Header() {
   const [showLinks, setShowLinks] = useState(false);
   const [showPopover, setShowPopover] = useState(false);
 
-  const { isLogged, setIsLogged } = useAuth();
-
   const handleShowLinks = () => setShowLinks(!showLinks);
-
-  const handleLogout = () => {
-    setIsLogged(false);
-    setShowPopover(false);
-  };
 
   const handleAvatarToggle = () => {
     setShowPopover((prev) => !prev);
@@ -64,53 +57,7 @@ function Header() {
               className="avatar-icon"
             />
           </button>
-
-          {showPopover && (
-            <div className="popover">
-              {isLogged ? (
-                <>
-                  <Link
-                    to="/app/account"
-                    onClick={() => setShowPopover(false)}
-                    className="popover-link"
-                  >
-                    Mon compte
-                  </Link>
-                  <Link
-                    to="/app/artist-profile"
-                    onClick={() => setShowPopover(false)}
-                    className="popover-link"
-                  >
-                    Mon profil artiste
-                  </Link>
-                  <button
-                    type="button"
-                    onClick={handleLogout}
-                    className="popover-button"
-                  >
-                    Se déconnecter
-                  </button>
-                </>
-              ) : (
-                <>
-                  <Link
-                    to="/login"
-                    onClick={() => setShowPopover(false)}
-                    className="popover-link"
-                  >
-                    Se connecter
-                  </Link>
-                  <Link
-                    to="/app/signin"
-                    onClick={() => setShowPopover(false)}
-                    className="popover-link"
-                  >
-                    Créer un compte
-                  </Link>
-                </>
-              )}
-            </div>
-          )}
+          <UserMenu showPopover={showPopover} setShowPopover={setShowPopover} />
 
           <button
             type="button"
