+++ client/src/types/userMenu.d.ts
@@ -1,0 +1,4 @@
+type UserMenuProps = {
+  showPopover: boolean;
+  setShowPopover: (boolean) => void;
+};
